autogear/requirements.txt
asgiref==3.8.1
configparser==7.0.0
Django==5.0.7
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
PyMySQL==1.1.1
sqlparse==0.5.1
